requirements.txt
alpaca-py>=0.33.0
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
pytz>=2024.1
requests>=2.31.0
